Mostly machine-convert to libpeas
A machine-generated .plugin.in is created for for each plugin. This is
then preprocessed by sed and intltool to generate the final .plugin file.

Additionally each plugin links with the new peas-support which adds the
necessary runtime support to enable libpeas to load it. peas-support.c
1) creates a GObject class wrapper that is instantiable by geany, and 2)
it exports peas_register_types() that libpeas requires (that is
the only global symbol required).

Plugins should be converted to work with libpeas natively, then they
can drop support library. peas-support.c should give maintainers enough
headroom to perform this task.